トップ画面の完成と不具合の修正
ウイスキーの分類が複数行になってしまう使用を修正。分類ごとの画面に対応。
